+++ src/main/java/ladder/domain/common/LadderMessage.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.common;
 
 public final class LadderMessage {
 
@@ -6,11 +6,11 @@ public final class LadderMessage {
 
     public static final String LINE_LENGTH_MESSAGE = "The line length must be greater than zero";
 
-    public static final String BLANK_NAME_MESSAGE = "The person name cannot be blank";
+    public static final String BLANK_NAME_MESSAGE = "The name cannot be blank";
 
-    public static final String NAME_LENGTH_MESSAGE = "The person name can only be %s characters or less";
+    public static final String NAME_LENGTH_MESSAGE = "The name can only be %s characters or less";
 
     public static final String BLANK_TEXT_MESSAGE = "The text cannot be blank";
 
-    public static final String LADDER_RESULT_MESSAGE = "The ladder result cannot be blank";
+    public static final String PERSON_PRIZE_SIZE_MESSAGE = "Person and result size are different";
 }
