+++ src/main/java/ladder/domain/common/Printable.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.common;
 
 public interface Printable {
     String print();
