+++ src/main/java/ladder/domain/line/DrawStrategy.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.line;
 
 import java.util.function.Consumer;
