+++ src/test/java/ladder/domain/line/RandomDrawStrategyTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.line;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
